Deepseek/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
python-dotenv
psycopg2-binary
numpy
pandas
setuptools
python-multipart